+++ src/Sulu/Bundle/RouteBundle/Tests/Unit/Routing/RouteProviderTest.php
@@ -83,6 +83,7 @@ class RouteProviderTest extends \PHPUnit_Framework_TestCase
         $request->getLocale()->willReturn('de');
 
         $routeEntity = $this->prophesize(RouteInterface::class);
+        $routeEntity->getId()->willReturn(1);
         $routeEntity->getEntityClass()->willReturn('Example');
         $routeEntity->isHistory()->willReturn(false);
 
@@ -136,6 +137,35 @@ class RouteProviderTest extends \PHPUnit_Framework_TestCase
         $this->defaultsProvider->isPublished('Example', '1', 'de')->willReturn(true);
         $this->defaultsProvider->getByEntity('Example', '1', 'de')->willReturn(['test' => 1]);
 
+        $collection = $this->routeProvider->getRouteCollectionForRequest($request->reveal());
+
+        $this->assertCount(1, $collection);
+        $routes = array_values(iterator_to_array($collection->getIterator()));
+
+        $this->assertEquals('/de/test', $routes[0]->getPath());
+        $this->assertEquals(['test' => 1], $routes[0]->getDefaults());
+    }
+
+    public function testGetRouteCollectionForRequestTwice()
+    {
+        $request = $this->prophesize(Request::class);
+        $request->getPathInfo()->willReturn('/de/test');
+        $request->getLocale()->willReturn('de');
+
+        $routeEntity = $this->prophesize(RouteInterface::class);
+        $routeEntity->getEntityClass()->willReturn('Example');
+        $routeEntity->getEntityId()->willReturn('1');
+        $routeEntity->getId()->willReturn(1);
+        $routeEntity->getPath()->willReturn('/test');
+        $routeEntity->isHistory()->willReturn(false);
+        $routeEntity->getLocale()->willReturn('de');
+
+        $this->routeRepository->findByPath('/test', 'de')->willReturn($routeEntity->reveal())->shouldBeCalledTimes(1);
+        $this->defaultsProvider->supports('Example')->willReturn(true)->shouldBeCalledTimes(1);
+        $this->defaultsProvider->isPublished('Example', '1', 'de')->willReturn(true)->shouldBeCalledTimes(1);
+        $this->defaultsProvider->getByEntity('Example', '1', 'de')->willReturn(['test' => 1])->shouldBeCalledTimes(1);
+
+        $this->routeProvider->getRouteCollectionForRequest($request->reveal());
         $collection = $this->routeProvider->getRouteCollectionForRequest($request->reveal());
 
         $this->assertCount(1, $collection);